Added user.go in api to transfer user group endpoints
+ user.go in api package
* /user/ endpoints to own file
